Revert "Merge pull request #8906 from remibettan/rs-convert-support-imu-pose"
This reverts commit aaebe18bb252560ba6f80c93dbcc2480aa7eeaeb, reversing
changes made to b505a6f3df6e4856000e955297c05ac67debe59a.
