+++ tools/convert/readme.md
@@ -9,11 +9,11 @@ Console app for converting ROS-bag files to various formats (currently supported
 |Flag   |Description   |Default|
 |---|---|---|
 |`-i <ros-bag-file>`|ROS-bag filename||
-|`-p <png-path>`|convert to PNG, set output path to png-path||
-|`-v <csv-path>`|convert to CSV, set output path to csv-path, supported formats: depth, color, imu, pose||
-|`-r <raw-path>`|convert to RAW, set output path to raw-path||
-|`-l <ply-path>`|convert to PLY, set output path to ply-path||
-|`-b <bin-path>`|convert to BIN (depth matrix), set output path to bin-path||
+|`-p <png-path>`|convert to PNG, set output path to <png-path>||
+|`-v <csv-path>`|convert to CSV (depth matrix), set output path to <csv-path>||
+|`-r <raw-path>`|convert to RAW, set output path to <raw-path>||
+|`-l <ply-path>`|convert to PLY, set output path to <ply-path>||
+|`-b <bin-path>`|convert to BIN (depth matrix), set output path to <bin-path>||
 |`-d`|convert depth frames only||
 |`-c`|convert color frames only||
